Feat: feature division method (close - #8)

+++ src/main/java/com/ohgiraffers/run/Application.java
@@ -10,13 +10,13 @@ public class Application {
 
         Scanner sc = new Scanner(System.in);
 
-        double num1, num2;
+        int num1, num2;
 
         System.out.println("첫번째 숫자 입력하세요 : ");
-        num1 = sc.nextDouble();
+        num1 = sc.nextInt();
 
         System.out.println("두번째 숫자 입력하세요 : ");
-        num2 = sc.nextDouble();
+        num2 = sc.nextInt();
 
         Division division = new Division();
         System.out.println(division.div(num1,num2));
